add profile picute to suggestion dropdown
