+++ .eslintcache
@@ -1,1 +1,1 @@
-[{"/home/godson/Documents/Microverse/React/github-typeahead/src/components/App.js":"1","/home/godson/Documents/Microverse/React/github-typeahead/src/components/typeaheadDropdown.js":"2","/home/godson/Documents/Microverse/React/github-typeahead/src/reportWebVitals.js":"3","/home/godson/Documents/Microverse/React/github-typeahead/src/index.js":"4","/home/godson/Documents/Microverse/React/github-typeahead/src/components/API.js":"5","/home/godson/Documents/Microverse/React/github-typeahead/src/components/demo.js":"6"},{"size":282,"mtime":1608320095940,"results":"7","hashOfConfig":"8"},{"size":411,"mtime":1608319206685,"results":"9","hashOfConfig":"8"},{"size":362,"mtime":1608153629808,"results":"10","hashOfConfig":"8"},{"size":511,"mtime":1608157134960,"results":"11","hashOfConfig":"8"},{"size":328,"mtime":1608311255378,"results":"12","hashOfConfig":"8"},{"size":1281,"mtime":1608470068291,"results":"13","hashOfConfig":"8"},{"filePath":"14","messages":"15","errorCount":0,"warningCount":1,"fixableErrorCount":0,"fixableWarningCount":0,"source":"16","usedDeprecatedRules":"17"},"1r44178",{"filePath":"18","messages":"19","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"20","messages":"21","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"22"},{"filePath":"23","messages":"24","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"25","messages":"26","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"27","messages":"28","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"/home/godson/Documents/Microverse/React/github-typeahead/src/components/App.js",["29"],"import React from 'react';\nimport '../style/App.css';\nimport TypeaheadDropdown from './typeaheadDropdown';\nimport Input from './demo';\n\nconst App = () => {\n return (\n   <div className=\"App\">\n       {/* <TypeaheadDropdown /> */}\n       <Input />\n   </div>\n );\n}\n \nexport default App;",["30","31"],"/home/godson/Documents/Microverse/React/github-typeahead/src/components/typeaheadDropdown.js",[],"/home/godson/Documents/Microverse/React/github-typeahead/src/reportWebVitals.js",[],["32","33"],"/home/godson/Documents/Microverse/React/github-typeahead/src/index.js",[],"/home/godson/Documents/Microverse/React/github-typeahead/src/components/API.js",[],"/home/godson/Documents/Microverse/React/github-typeahead/src/components/demo.js",[],{"ruleId":"34","severity":1,"message":"35","line":3,"column":8,"nodeType":"36","messageId":"37","endLine":3,"endColumn":25},{"ruleId":"38","replacedBy":"39"},{"ruleId":"40","replacedBy":"41"},{"ruleId":"38","replacedBy":"42"},{"ruleId":"40","replacedBy":"43"},"no-unused-vars","'TypeaheadDropdown' is defined but never used.","Identifier","unusedVar","no-native-reassign",["44"],"no-negated-in-lhs",["45"],["44"],["45"],"no-global-assign","no-unsafe-negation"]
+[{"/home/godson/Documents/Microverse/React/github-typeahead/src/components/App.js":"1","/home/godson/Documents/Microverse/React/github-typeahead/src/components/typeaheadDropdown.js":"2","/home/godson/Documents/Microverse/React/github-typeahead/src/reportWebVitals.js":"3","/home/godson/Documents/Microverse/React/github-typeahead/src/index.js":"4","/home/godson/Documents/Microverse/React/github-typeahead/src/components/API.js":"5"},{"size":229,"mtime":1608499493210,"results":"6","hashOfConfig":"7"},{"size":1706,"mtime":1608501283793,"results":"8","hashOfConfig":"7"},{"size":362,"mtime":1608153629808,"results":"9","hashOfConfig":"7"},{"size":511,"mtime":1608157134960,"results":"10","hashOfConfig":"7"},{"size":328,"mtime":1608311255378,"results":"11","hashOfConfig":"7"},{"filePath":"12","messages":"13","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"1r44178",{"filePath":"14","messages":"15","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"16","messages":"17","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"18"},{"filePath":"19","messages":"20","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"21","messages":"22","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"/home/godson/Documents/Microverse/React/github-typeahead/src/components/App.js",[],"/home/godson/Documents/Microverse/React/github-typeahead/src/components/typeaheadDropdown.js",[],"/home/godson/Documents/Microverse/React/github-typeahead/src/reportWebVitals.js",[],["23","24"],"/home/godson/Documents/Microverse/React/github-typeahead/src/index.js",[],"/home/godson/Documents/Microverse/React/github-typeahead/src/components/API.js",[],{"ruleId":"25","replacedBy":"26"},{"ruleId":"27","replacedBy":"28"},"no-native-reassign",["29"],"no-negated-in-lhs",["30"],"no-global-assign","no-unsafe-negation"]
