+++ src/components/App.js
@@ -1,13 +1,11 @@
 import React from 'react';
 import '../style/App.css';
 import TypeaheadDropdown from './typeaheadDropdown';
-import Input from './demo';
 
 const App = () => {
  return (
    <div className="App">
-       {/* <TypeaheadDropdown /> */}
-       <Input />
+       <TypeaheadDropdown />
    </div>
  );
 }
