+++ src/components/typeaheadDropdown.js
@@ -1,19 +1,60 @@
-import React from 'react';
+import React, { useState } from 'react';
+import useUser from './API';
 import '../style/typeaheadDropdown.css';
 
-const TypeAheadDropDown = () => {
-  const onTextChange = (e) => {
-    const value = e.target.value;
-    if (value.length > 0) {
-      return value;
-    }
+const Input = ({ u }) => {
+  const { user, isLoading, isError } = useUser(u);
+
+  //let text='';
+  let [text, setText] = useState('');
+  //let suggestions = [];
+  let [suggestions, setSuggestions] = useState([]);
+  
+  console.log(text,suggestions);
+  if (isError) return <div>failed to load</div>
+  if (isLoading) return <div>loading...</div>
+
+  const SuggestionSelected=(name)=>{
+    setText(name);
+    setSuggestions([]);
   }
+
+  const OnTextChange = (e) => {
+    text = e.target.value;
+    if (text.length > 0) {
+      let regex = new RegExp(`^${text}`, `i`);
+      suggestions = user.sort().filter(v => regex.test(v.login));
+    }
+    setText(text);
+    setSuggestions(suggestions);
+  }
+
+  const RenderSuggestions = () => {
+    if (suggestions.length > 0 && text.length > 0) {
+      return (
+        // <ul>
+        //   {suggestions.map(user => <li key={user.id} onClick={(e)=>SuggestionSelected(user.login)}>{user.login}</li>)}
+        // </ul>
+        <div className="suggestions">
+        {suggestions.map(user => 
+        <div className="grid-container" key={user.id} onClick={(e)=>SuggestionSelected(user.login)}>
+            <div className="avatar"><img src={user.avatar_url} alt="Avatar"/></div>
+            <div className="name">
+              <p>{user.login}</p>
+            </div>
+        </div>
+        )}
+        </div>
+        
+      )
+    }else return null;
+  } 
  
   return (
     <div className="TypeAheadDropDown">
-      <input onChange={onTextChange} placeholder="Search user" type="text" />
+      <input onChange={OnTextChange} placeholder="Search user"  value={text} type="text" />
+      {RenderSuggestions()}
     </div>
-  );
- 
+  );  
 }
-export default TypeAheadDropDown;
+export default Input;
