+++ src/style/typeaheadDropdown.css
@@ -3,7 +3,7 @@
   border: 1px solid gray;
   box-shadow: 0 0 1px rgba(0,0,0,0.1), 0 2px 4px 1px rgba(0,0,0, .18);
   font-family: Arial, Helvetica, sans-serif;
-  font-size: 14px;
+  font-size: 62.5%;
   color: rgba(0, 0, 0, 0.73);
 }
   
@@ -11,18 +11,18 @@
   width: 100%;
   border: none;
   font-family: Arial, Helvetica, sans-serif;
-  font-size: 14px;
+  font-size: 2rem;
   color: rgba(0, 0, 0, 0.73);
-  padding: 10px 5px ;
+  padding: 1rem 0.5rem ;
   box-sizing: border-box;
   outline: none;
 
 }
-.TypeAheadDropDown ul::before{
+.TypeAheadDropDown .suggestions::before{
   content: '';
 }
 
-.TypeAheadDropDown ul{
+.TypeAheadDropDown .suggestions{
   list-style-type: none;
   text-align: left;
   margin: 0;
@@ -30,17 +30,32 @@
   border-top: 1px solid gray;
 }
 
-.TypeAheadDropDown li{
-  padding: 10px 5px ;
+.TypeAheadDropDown .grid-container{
+  padding: 1rem 0.5rem ;
   cursor: pointer;
 }
 
-.TypeAheadDropDown li:hover{
+.TypeAheadDropDown .grid-container:hover{
   background: lightgray;
   text-decoration: underline;
 }
 
+.TypeAheadDropDown .grid-container img{
+  display : flex;
+}
+
+.TypeAheadDropDown .grid-container .avatar{
+  height: 4rem;
+  width: auto;
+}
+
+.TypeAheadDropDown .grid-container img{
+  height: 100%;
+  width: auto;
+  border-radius: 50%;
+}
+
 .App{
   margin: 150px auto 0 auto;
-  width: 600px;
+  width: 60rem;
 }
